refactor: audit and harden 58 files across SDK (R132-R231)
Systematic code quality audit covering unwrap elimination, clippy fixes,
type safety improvements, and dead code removal across all major modules.

Key changes:
- Eliminate unsafe .unwrap()/.expect() in production code paths
- Fix clone_on_copy: add Copy to CacheConfig, remove redundant .clone()
- Type safety: i32→i64 for neo_token votes/gas, u64 as i64→try_from
- Fix PartialEq bug in ProviderError (7 missing match arms)
- Fix unreachable match arms in invocation_result deserializer
- Replace expect_throw with recoverable error in wasm ws/backend
- Simplify invoke_function match block, remove needless return
- Use matches!() macro, .values() iterator, into_owned() idioms
- Remove unused imports in monitoring/metrics and monitoring/tracing
- Add Default impl for SecureStorage, fix non-existent SgxError variant
- Safer float-to-u64 ceiling, lossless u32→u64 via From trait

+++ src/neo_protocol/responses/neo_block.rs
@@ -23,9 +23,9 @@ pub struct NeoBlock {
 	#[serde(deserialize_with = "deserialize_h256")]
 	#[serde(rename = "merkleroot")]
 	pub merkle_root_hash: H256,
-	pub time: i32,
+	pub time: u64,
 	pub nonce: String,
-	pub index: i32,
+	pub index: u32,
 	pub primary: Option<i32>,
 	#[serde(rename = "nextconsensus")]
 	pub next_consensus: String,
